pyproject.toml
[tool.poetry]
name = "shopback"
version = "0.1.0"
description = ""
authors = ["davidtnfsh <davidtnfsh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
UdicOpenData = "^2.4"
sklearn = "^0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
mypy = "^0.770"
black = "^19.10b0"
pytest-cov = "^2.8.1"
isort = "^4.3.21"
flake8 = "^3.7.9"
pre-commit = "^2.2.0"
commitizen = "^1.17.0"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.1"
tag_format = "$version"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
